chore: make bench a script
