+++ .npmignore
@@ -1,3 +1,4 @@
-test/
+bin/bench.js
 examples/
+test/
 .travis.yml
