+++ package.json
@@ -11,13 +11,14 @@
   "homepage": "https://github.com/mafintosh/csv-parser",
   "bugs": "https://github.com/mafintosh/csv-parser/issues",
   "bin": {
-    "csv-parser": "./bin.js"
+    "csv-parser": "./bin/csv-parser.js"
   },
   "main": "index.js",
   "engines": {
     "node": ">= 6.14.0"
   },
   "scripts": {
+    "bench": "find test/data -maxdepth 1 -type f -exec scripts/bench",
     "commitlint": "commitlint",
     "commitmsg": "commitlint -e $GIT_PARAMS",
     "lint": "eslint .",
